+++ pkg/accounts/service/authenticationTokenService.test.ts
@@ -2,7 +2,7 @@ import { Option, Result } from '@mikuroxina/mini-fn';
 import * as jose from 'jose';
 import { describe, expect, it } from 'vitest';
 
-import { MockClock } from '../../id/mod.js';
+import { MockClock } from '../../internal/id/mod.js';
 import { AuthenticationTokenService } from './authenticationTokenService.js';
 
 describe('AuthenticationTokenService', () => {
