+++ pkg/accounts/service/authenticationTokenService.ts
@@ -1,7 +1,7 @@
 import { Ether, Option, type Promise, Result } from '@mikuroxina/mini-fn';
 import * as jose from 'jose';
 
-import { type Clock, clockSymbol } from '../../id/mod.js';
+import { type Clock, clockSymbol } from '../../internal/id/mod.js';
 import {
   AccountAuthenticationTokenExpiredError,
   AccountAuthenticationTokenInvalidError,
