+++ pkg/accounts/service/follow.test.ts
@@ -1,7 +1,6 @@
 import { Option, Result } from '@mikuroxina/mini-fn';
 import { describe, expect, it } from 'vitest';
-
-import { MockClock } from '../../id/mod.js';
+import { MockClock } from '../../internal/id/mod.js';
 import { InMemoryAccountRepository } from '../adaptor/repository/dummy/account.js';
 import { InMemoryAccountFollowRepository } from '../adaptor/repository/dummy/follow.js';
 import { Account, type AccountID } from '../model/account.js';
