+++ pkg/accounts/service/follow.ts
@@ -1,6 +1,6 @@
 import { Ether, Option, Result } from '@mikuroxina/mini-fn';
 
-import { type Clock, clockSymbol } from '../../id/mod.js';
+import { type Clock, clockSymbol } from '../../internal/id/mod.js';
 import type { AccountName } from '../model/account.js';
 import { AccountNotFoundError } from '../model/errors.js';
 import { AccountFollow } from '../model/follow.js';
