+++ pkg/accounts/service/resendToken.test.ts
@@ -1,8 +1,7 @@
 import { Option } from '@mikuroxina/mini-fn';
 import { afterEach, describe, expect, it } from 'vitest';
-
-import { MockClock } from '../../id/mod.js';
 import { notificationModule } from '../../intermodule/notification.js';
+import { MockClock } from '../../internal/id/mod.js';
 import { InMemoryAccountRepository } from '../adaptor/repository/dummy/account.js';
 import { InMemoryAccountVerifyTokenRepository } from '../adaptor/repository/dummy/verifyToken.js';
 import { Account, type AccountID } from '../model/account.js';
