+++ pkg/drive/service/upload.test.ts
@@ -3,7 +3,7 @@ import { Result } from '@mikuroxina/mini-fn';
 import { describe, expect, it } from 'vitest';
 
 import type { AccountID } from '../../accounts/model/account.js';
-import { MockClock, SnowflakeIDGenerator } from '../../id/mod.js';
+import { MockClock, SnowflakeIDGenerator } from '../../internal/id/mod.js';
 import { InMemoryMediaRepository } from '../adaptor/repository/dummy.js';
 import { LocalStorage } from '../adaptor/storage/dummy.js';
 import {
