+++ pkg/notification/model/notificationBase.ts
@@ -1,6 +1,6 @@
 import { Option } from '@mikuroxina/mini-fn';
 import type { AccountID } from '../../accounts/model/account.js';
-import type { ID } from '../../id/type.js';
+import type { ID } from '../../internal/id/type.js';
 
 export type NotificationID = ID<NotificationBase>;
 export type NotificationType =
