+++ pkg/notification/service/create.test.ts
@@ -1,7 +1,7 @@
 import { Result } from '@mikuroxina/mini-fn';
 import { describe, expect, it } from 'vitest';
 import type { AccountID } from '../../accounts/model/account.js';
-import { MockClock, SnowflakeIDGenerator } from '../../id/mod.js';
+import { MockClock, SnowflakeIDGenerator } from '../../internal/id/mod.js';
 import type { NoteID } from '../../notes/model/note.js';
 import type { ReactionID } from '../../notes/model/reaction.js';
 import { InMemoryNotificationRepository } from '../adaptor/repository/dummy/notification.js';
