+++ pkg/notification/service/markAsRead.ts
@@ -1,6 +1,6 @@
 import { Ether, Result } from '@mikuroxina/mini-fn';
 import type { AccountID } from '../../accounts/model/account.js';
-import { type Clock, clockSymbol } from '../../id/mod.js';
+import { type Clock, clockSymbol } from '../../internal/id/mod.js';
 import type {
   Notification,
   NotificationID,
