+++ pkg/timeline/service/createList.test.ts
@@ -2,7 +2,7 @@ import { Result } from '@mikuroxina/mini-fn';
 import { describe, expect, it } from 'vitest';
 
 import type { AccountID } from '../../accounts/model/account.js';
-import { MockClock, SnowflakeIDGenerator } from '../../id/mod.js';
+import { MockClock, SnowflakeIDGenerator } from '../../internal/id/mod.js';
 import { InMemoryListRepository } from '../adaptor/repository/dummy.js';
 import { CreateListService } from './createList.js';
