+++ pkg/timeline/service/push.test.ts
@@ -4,8 +4,8 @@ import { beforeEach, describe, expect, it, vi } from 'vitest';
 import type { AccountID } from '../../accounts/model/account.js';
 import { partialAccount1 } from '../../accounts/testData/testData.js';
 import { dummyAccountModuleFacade } from '../../intermodule/account.js';
+import { addSecondsToDate } from '../../internal/time/mod.js';
 import { Note, type NoteID } from '../../notes/model/note.js';
-import { addSecondsToDate } from '../../time/mod.js';
 import { InMemoryListRepository } from '../adaptor/repository/dummy.js';
 import { InMemoryTimelineCacheRepository } from '../adaptor/repository/dummyCache.js';
 import { List, type ListID } from '../model/list.js';
